feat(prediction)!: 예상 질문 저장 Command로 바인딩하는 로직을 별도의 클래스로 분리한다.
OpenAi와 관련된, ChatCompletion과 관련된 클래스로 바인딩하는 로직만을 담당하도록 한다.
