+++ resumarble-core/src/main/kotlin/resumarble/core/domain/prediction/mapper/PredictionMapper.kt
@@ -1,7 +1,5 @@
-package resumarble.core.domain.gpt
+package resumarble.core.domain.prediction.mapper
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
 import resumarble.core.domain.prediction.domain.Answer
 import resumarble.core.domain.prediction.domain.Question
 import resumarble.core.domain.prediction.domain.QuestionAndAnswer
@@ -10,37 +8,10 @@ import resumarble.core.domain.prediction.domain.constraints.Job
 import resumarble.core.domain.prediction.facade.SavePredictionCommand
 import resumarble.core.domain.resume.facade.InterviewQuestion
 import resumarble.core.domain.resume.facade.InterviewQuestionCommand
-import resumarble.core.domain.resume.facade.InterviewQuestionResponse
 import resumarble.core.global.annotation.Mapper
 
 @Mapper
-class OpenAiMapper(
-    private val objectMapper: ObjectMapper
-) {
-    fun promptAndContentToChatCompletionRequest(
-        prompt: String,
-        content: String
-    ): ChatCompletionRequest {
-        return ChatCompletionRequest(
-            messages = listOf(
-                ChatCompletionMessage(
-                    role = OpenAiRole.SYSTEM.value,
-                    content = prompt
-                ),
-                ChatCompletionMessage(
-                    role = OpenAiRole.USER.value,
-                    content = content
-                )
-            )
-        )
-    }
-
-    fun completionToInterviewQuestionResponse(completion: ChatCompletionMessageResponse): List<InterviewQuestion> {
-        val interviewQuestionResponse = objectMapper.readValue<InterviewQuestionResponse>(
-            completion.questionAndAnswer
-        )
-        return interviewQuestionResponse.interviews
-    }
+object PredictionMapper {
 
     fun completionToSavePredictionCommand(
         command: InterviewQuestionCommand,
